`SHs`: move `SType` into its own module
