+++ hs-bindgen/src-internal/HsBindgen/Backend/SHs/AST.hs
@@ -48,7 +48,7 @@ import HsBindgen.Backend.Hs.Haddock.Documentation qualified as HsDoc
 import HsBindgen.Backend.Hs.Name qualified as Hs
 import HsBindgen.Backend.Hs.Origin qualified as Origin
 import HsBindgen.Backend.Level
-import HsBindgen.BindingSpec qualified as BindingSpec
+import HsBindgen.Backend.SHs.AST.Type
 import HsBindgen.Frontend.Naming
 import HsBindgen.Imports
 import HsBindgen.Instances qualified as Inst
@@ -60,13 +60,6 @@ import HsBindgen.NameHint
 -------------------------------------------------------------------------------}
 
 type ClosedExpr = SExpr EmptyCtx
-
--- | Natural number larger equal 2.
-newtype Plus2 = Plus2 Natural
-  deriving stock (Show, Eq, Ord)
-
-applyPlus2 :: Plus2 -> Natural
-applyPlus2 (Plus2 n) = n+2
 
 -- | Simple expressions
 type SExpr :: Ctx -> Star
@@ -173,35 +166,11 @@ data SDecl =
   | DPatternSynonym PatternSynonym
   deriving stock (Show)
 
-type ClosedType = SType EmptyCtx
-
--- | Simple types
-type SType :: Ctx -> Star
-data SType ctx =
-    TGlobal (Global LvlType)
-  | TClass Inst.TypeClass
-  | TCon (Hs.Name Hs.NsTypeConstr)
-  | TFun (SType ctx) (SType ctx)
-  | TLit Natural
-  | TStrLit String
-  | TExt Hs.ExtRef BindingSpec.CTypeSpec BindingSpec.HsTypeSpec
-  | TBound (Idx ctx)
-  | TFree (Hs.Name Hs.NsVar)
-  | TApp (SType ctx) (SType ctx)
-  | TUnit
-  | TBoxedTup Plus2
-  | TEq
-  | forall n ctx'. TForall (Vec n NameHint) (Add n ctx ctx') [SType ctx'] (SType ctx')
-
 tBindgenGlobal :: BindgenGlobalType -> SType ctx
 tBindgenGlobal = TGlobal . bindgenGlobalType
 
 data Pragma = NOINLINE
   deriving stock Show
-
-infixl 9 `TApp`
-
-deriving stock instance Show (SType ctx)
 
 data TypeSynonym = TypeSynonym{
       name    :: Hs.Name Hs.NsTypeConstr
